* implement base16 colorscheme for mangowc * add mangowc waybar config * added ability for vol.sh to refresh waybar

+++ .local/scripts/vol.sh
@@ -3,7 +3,13 @@
 PERC="0.05"
 
 refbar() {
-    pkill sleep -P "$(cat ~/.cache/statusbar_pid )"
+    if [ "$(pgrep swaybar)" ]; then
+        pkill sleep -P "$(cat ~/.cache/statusbar_pid )"
+    fi
+
+    if [ "(pgrep waybar)" ]; then
+        kill -n 37 $(pgrep waybar)
+    fi
 }
 
 display() {
